Add Nuspec Schema scrubber and add to default scrubbers list (#11)
The NuGet client will always use the oldest schema version it can.
Additionally, a new schema version was released for SemVer version
numbers (i.e. v2 and v3 here:
https://github.com/NuGet/NuGet.Client/blob/8c972cdff5b1194d7c37384fca5816a33ffbe0c4/src/NuGet.Core/NuGet.Packaging/PackageCreation/Authoring/ManifestSchemaUtility.cs).

The result of these two factors is that a version bump (such as from
1.0.0-deadbeef to 1.0.1) may result in a nuspec diff.
Thus scrub the xmlns for the NuSpec schema to reduce the noise in the
diff.