Reduce diff noise in ASCII tree (#12)
Extend the folder bars down for all files instead of only for files that
aren't the last child. This reduces the amount of noise in diffs when
files are added / removed, as bars no longer need to be changed.

This change also removes the dependency on Spectre.Console, which was
only used for diff rendering.