Replace NupkgFinder / ArtifactsPath embedding with GetPackFromProject (#13)
Embedding the ArtifactsPath into the test assembly _works_, but I
dislike it because it makes the test projects not reproducible /
non-deterministic.

Instead use GetPackFromProject to copy the .nupkg from the project's
output into the test project's output and use it from there. This
pattern both follows build determinisim rules better, and is "safer" in
that it prevents the tests from accessing the main package, instead
using their own private copy.

Also update README to reference GetPackFromProject so consumers can use
the same technique.